+++ src/compas/geometry/_core/angles.py
@@ -86,9 +86,8 @@ def angle_vectors_signed(u, v, normal, deg=False, threshold=1e-3):
     normal_uv = cross_vectors(u, v)
 
     if length_vector(normal_uv) > threshold:
-        # check if normal_uv has the same direction as normal
-        angle_btw_normals = angle_vectors(normal, normal_uv)
-        if angle_btw_normals > threshold:
+        # check if normal_uv has the same direction as normals
+        if dot_vectors(normal, normal_uv) < -threshold:
             angle *= -1
 
     if deg: